Minor fixes after review

+++ apps/backend/src/main/java/no/nav/data/etterlevelse/krav/KravService.java
@@ -170,6 +170,9 @@ public class KravService {
     @Transactional(propagation = Propagation.REQUIRED)
     public Krav delete(UUID id) {
         var kravDelete = get(id);
+        if (kravDelete == null) {
+            return null;
+        }
         repo.deleteById(id);
         return kravDelete;
     }
